backend/requirements/base.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
psycopg2-binary>=2.9.9
redis>=5.0.0
gunicorn>=21.2.0
celery>=5.3.0
django-redis>=5.3.0
django-cors-headers>=4.2.0
django-filter>=23.2
Pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.8.5
python-dotenv>=1.0.0
